refactor(infraestrutura): Adicionando as implementações dos microserviços inventory-service e product-service no mesmo repositorio
