+++ inventory-service/src/main/java/com/viniccius/inventoryservice/InventoryServiceApplication.java
@@ -1,0 +1,13 @@
+package com.viniccius.inventoryservice;
+
+import org.springframework.boot.SpringApplication;
+import org.springframework.boot.autoconfigure.SpringBootApplication;
+
+@SpringBootApplication
+public class InventoryServiceApplication {
+
+	public static void main(String[] args) {
+		SpringApplication.run(InventoryServiceApplication.class, args);
+	}
+
+}
