+++ inventory-service/src/main/resources/application.properties
@@ -1,0 +1,2 @@
+spring.application.name=inventory-service
+server.port=9090
