+++ inventory-service/src/test/java/com/viniccius/inventoryservice/InventoryServiceApplicationTests.java
@@ -1,0 +1,13 @@
+package com.viniccius.inventoryservice;
+
+import org.junit.jupiter.api.Test;
+import org.springframework.boot.test.context.SpringBootTest;
+
+@SpringBootTest
+class InventoryServiceApplicationTests {
+
+	@Test
+	void contextLoads() {
+	}
+
+}
