+++ product-service/src/main/java/com/viniccius/aws/productservice/ProductServiceApplication.java
@@ -1,0 +1,13 @@
+package com.viniccius.aws.productservice;
+
+import org.springframework.boot.SpringApplication;
+import org.springframework.boot.autoconfigure.SpringBootApplication;
+
+@SpringBootApplication
+public class ProductServiceApplication {
+
+	public static void main(String[] args) {
+		SpringApplication.run(ProductServiceApplication.class, args);
+	}
+
+}
