+++ product-service/src/main/java/com/viniccius/aws/productservice/dto/request/ProductRequestDTO.java
@@ -1,0 +1,10 @@
+package com.viniccius.aws.productservice.dto.request;
+
+public record ProductRequestDTO(
+        Long id,
+        String name,
+        String model,
+        String code,
+        float price,
+        String color
+) {}
