+++ product-service/src/main/java/com/viniccius/aws/productservice/dto/response/ProductResponseDTO.java
@@ -1,0 +1,10 @@
+package com.viniccius.aws.productservice.dto.response;
+
+public record ProductResponseDTO(
+        Long id,
+        String name,
+        String model,
+        String code,
+        float price,
+        String color
+) {}
