+++ product-service/src/main/java/com/viniccius/aws/productservice/enums/EventType.java
@@ -1,0 +1,7 @@
+package com.viniccius.aws.productservice.enums;
+
+public enum EventType {
+    PRODUCT_CREATED,
+    PRODUCT_UPDATED,
+    PRODUCT_DELETED
+}
