+++ product-service/src/main/java/com/viniccius/aws/productservice/exception/ProductNotFoundException.java
@@ -1,0 +1,7 @@
+package com.viniccius.aws.productservice.exception;
+
+public class ProductNotFoundException extends RuntimeException {
+    public ProductNotFoundException(String message) {
+        super(message);
+    }
+}
