+++ product-service/src/main/java/com/viniccius/aws/productservice/model/Envelope.java
@@ -1,0 +1,10 @@
+package com.viniccius.aws.productservice.model;
+
+import com.viniccius.aws.productservice.enums.EventType;
+import lombok.Data;
+
+@Data
+public class Envelope {
+    private EventType eventType;
+    private String data;
+}
