+++ product-service/src/main/java/com/viniccius/aws/productservice/model/Product.java
@@ -1,0 +1,38 @@
+package com.viniccius.aws.productservice.model;
+
+import jakarta.persistence.*;
+import lombok.AllArgsConstructor;
+import lombok.Builder;
+import lombok.Data;
+import lombok.NoArgsConstructor;
+
+@Table(
+        uniqueConstraints = {
+                @UniqueConstraint(columnNames = {"code"})
+        }
+)
+@Entity
+@Data
+@Builder
+@NoArgsConstructor
+@AllArgsConstructor
+public class Product {
+
+    @Id
+    @GeneratedValue(strategy = GenerationType.IDENTITY)
+    private Long id;
+
+    @Column(length = 32, nullable = false)
+    private String name;
+
+    @Column(length = 24, nullable = false)
+    private String model;
+
+    @Column(length = 8, nullable = false)
+    private String code;
+
+    private float price;
+
+    @Column(length = 12)
+    private String color;
+}
