+++ product-service/src/main/java/com/viniccius/aws/productservice/model/ProductEvent.java
@@ -1,0 +1,10 @@
+package com.viniccius.aws.productservice.model;
+
+import lombok.Data;
+
+@Data
+public class ProductEvent {
+    private Long productId;
+    private String code;
+    private String username;
+}
