+++ product-service/src/main/java/com/viniccius/aws/productservice/repository/ProductRepository.java
@@ -1,0 +1,10 @@
+package com.viniccius.aws.productservice.repository;
+
+import com.viniccius.aws.productservice.model.Product;
+import org.springframework.data.jpa.repository.JpaRepository;
+
+import java.util.Optional;
+
+public interface ProductRepository extends JpaRepository<Product, Long> {
+    Optional<Product> findByCode(String code);
+}
