+++ product-service/src/main/java/com/viniccius/aws/productservice/service/ProductPublisher.java
@@ -1,0 +1,53 @@
+package com.viniccius.aws.productservice.service;
+
+import com.amazonaws.services.sns.AmazonSNS;
+import com.amazonaws.services.sns.model.Topic;
+import com.fasterxml.jackson.core.JsonProcessingException;
+import com.fasterxml.jackson.databind.ObjectMapper;
+import com.viniccius.aws.productservice.enums.EventType;
+import com.viniccius.aws.productservice.model.Envelope;
+import com.viniccius.aws.productservice.model.Product;
+import com.viniccius.aws.productservice.model.ProductEvent;
+import org.slf4j.Logger;
+import org.slf4j.LoggerFactory;
+import org.springframework.beans.factory.annotation.Qualifier;
+import org.springframework.stereotype.Service;
+
+@Service
+public class ProductPublisher {
+
+    private static final Logger logger = LoggerFactory.getLogger(ProductPublisher.class);
+    private final AmazonSNS snsClient;
+    private final Topic productEventsTopic;
+    private final ObjectMapper objectMapper;
+
+    public ProductPublisher(AmazonSNS snsClient,
+                            @Qualifier("productEventsTopic")Topic productEventsTopic,
+                            ObjectMapper objectMapper) {
+
+        this.snsClient = snsClient;
+        this.productEventsTopic = productEventsTopic;
+        this.objectMapper = objectMapper;
+    }
+
+    public void publishProductEvent(Product product, EventType eventType, String username) {
+        ProductEvent productEvent = new ProductEvent();
+        productEvent.setProductId(product.getId());
+        productEvent.setCode(product.getCode());
+        productEvent.setUsername(username);
+
+        Envelope envelope = new Envelope();
+        envelope.setEventType(eventType);
+        try {
+            envelope.setData(objectMapper.writeValueAsString(productEvent));
+
+            snsClient.publish(
+                    productEventsTopic.getTopicArn(),
+                    objectMapper.writeValueAsString(envelope)
+            );
+
+        } catch (JsonProcessingException e) {
+            throw new RuntimeException(e);
+        }
+    }
+}
