+++ product-service/src/main/resources/application.properties
@@ -1,0 +1,9 @@
+spring.application.name=product-service
+spring.jpa.hibernate.ddl-auto=update
+
+spring.datasource.url=jdbc:mariadb://localhost:3306/product_service?createDatabaseIfNotExist=true
+spring.datasource.username=root
+spring.datasource.password=
+
+aws.region=us-east-1
+aws.sns.topic.product.events.arn=product-events
