+++ product-service/src/test/java/com/viniccius/aws/productservice/ProductServiceApplicationTest.java
@@ -1,0 +1,12 @@
+package com.viniccius.aws.productservice;
+
+import org.junit.jupiter.api.Test;
+import org.springframework.boot.test.context.SpringBootTest;
+
+@SpringBootTest
+class ProductServiceApplicationTest {
+
+    @Test
+    void contextLoads() {
+    }
+}
