--- src/main/java/com/viniccius/aws/spring_aws_microservices/SpringAwsMicroservicesApplication.java
@@ -1,13 +1,0 @@
-package com.viniccius.aws.spring_aws_microservices;
-
-import org.springframework.boot.SpringApplication;
-import org.springframework.boot.autoconfigure.SpringBootApplication;
-
-@SpringBootApplication
-public class SpringAwsMicroservicesApplication {
-
-	public static void main(String[] args) {
-		SpringApplication.run(SpringAwsMicroservicesApplication.class, args);
-	}
-
-}
