--- src/main/java/com/viniccius/aws/spring_aws_microservices/dto/request/ProductRequestDTO.java
@@ -1,10 +1,0 @@
-package com.viniccius.aws.spring_aws_microservices.dto.request;
-
-public record ProductRequestDTO(
-        Long id,
-        String name,
-        String model,
-        String code,
-        float price,
-        String color
-) {}
