--- src/main/java/com/viniccius/aws/spring_aws_microservices/enums/EventType.java
@@ -1,7 +1,0 @@
-package com.viniccius.aws.spring_aws_microservices.enums;
-
-public enum EventType {
-    PRODUCT_CREATED,
-    PRODUCT_UPDATED,
-    PRODUCT_DELETED
-}
