--- src/main/java/com/viniccius/aws/spring_aws_microservices/exception/ProductNotFoundException.java
@@ -1,7 +1,0 @@
-package com.viniccius.aws.spring_aws_microservices.exception;
-
-public class ProductNotFoundException extends RuntimeException {
-    public ProductNotFoundException(String message) {
-        super(message);
-    }
-}
