--- src/main/resources/application.properties
@@ -1,9 +1,0 @@
-spring.application.name=spring-aws-microservices
-spring.jpa.hibernate.ddl-auto=update
-
-spring.datasource.url=jdbc:mariadb://localhost:3306/aws_microservice01?createDatabaseIfNotExist=true
-spring.datasource.username=root
-spring.datasource.password=
-
-aws.region=us-east-1
-aws.sns.topic.product.events.arn=product-events
