--- src/test/java/com/viniccius/aws/spring_aws_microservices/SpringAwsMicroservicesApplicationTests.java
@@ -1,13 +1,0 @@
-package com.viniccius.aws.spring_aws_microservices;
-
-import org.junit.jupiter.api.Test;
-import org.springframework.boot.test.context.SpringBootTest;
-
-@SpringBootTest
-class SpringAwsMicroservicesApplicationTests {
-
-	@Test
-	void contextLoads() {
-	}
-
-}
